+++ Rubeus/Domain/CommandCollection.cs
@@ -44,6 +44,7 @@ namespace Rubeus.Domain
             _availableCommands.Add(Silver.CommandName, () => new Silver());
             _availableCommands.Add(Golden.CommandName, () => new Golden());
             _availableCommands.Add(Diamond.CommandName, () => new Diamond());
+            _availableCommands.Add(Preauthscan.CommandName, () => new Preauthscan());
         }
 
         public bool ExecuteCommand(string commandName, Dictionary<string, string> arguments)