+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0]
+
+### Added
+
+* `preauthscan` command to scan for accounts that do not require Kerberos pre-authentication
+* `/preauth` argument to the `kerberoast` command, to kerberoast with an account that does not require Kerberos pre-authentication
+* `/nopreauth` flag to the `asktgt` command, to request a TGT without providing pre-authentication
+* `/service` argument to the `asktgt` command, to request service tickets using an AS-REQ
+
 ## [2.1.0]
 
 ### Added